Design: 모달 댓글 리스트 스타일 수정
