+++ front/components/Modal/ModalContent.tsx
@@ -6,33 +6,18 @@ import {
   DeleteOutlined,
   EditOutlined,
   LikeOutlined,
-  LoadingOutlined,
   MoreOutlined,
-  PaperClipOutlined,
-  SendOutlined,
-  ShareAltOutlined,
-  SmileOutlined
+  ShareAltOutlined
 } from '@ant-design/icons';
 
-import useInput from 'utils/useInput';
-import useFileUpload from 'utils/useFileUpload';
 import ModalCommentList from './ModalCommentList';
 import { RootState } from 'store/reducers';
 import { formatDate } from 'utils/useListTimes';
-import {
-  executePostEdit,
-  modalCommentRemoveUploadedImage,
-  modalCommentUploadImageRequest,
-  showDeleteModal
-} from 'store/actions/postAction';
+import { executePostEdit, modalCommentRemoveUploadedImage, showDeleteModal } from 'store/actions/postAction';
 
-import { slideInTooltip, slideInUploadImage } from 'styles/Common/animation';
+import { slideInTooltip } from 'styles/Common/animation';
 import { Tooltip, TooltipBtn, TooltipOutsideArea } from 'styles/Common/tooltip';
 import {
-  ModalCommentInput,
-  ModalCommentInputImage,
-  ModalCommentInputImageWrapper,
-  ModalCommentInputWrapper,
   ModalContentHeader,
   ModalContentOptions,
   ModalContentText,
@@ -42,18 +27,9 @@ import {
 const ModalContent = () => {
   const dispatch = useDispatch();
   const { me } = useSelector((state: RootState) => state.user);
-  const { singlePost, modalCommentImagePath, modalCommentUploadImageLoading } = useSelector(
-    (state: RootState) => state.post
-  );
-
-  const [comment, onChangeComment] = useInput('');
-  const { fileInputRef, onFileChange } = useFileUpload(modalCommentUploadImageRequest, { showWarning: false });
+  const { singlePost, modalCommentImagePath } = useSelector((state: RootState) => state.post);
   const [isModalCommentListVisible, setIsModalCommentListVisible] = useState<boolean>(false);
   const [isTooltipVisible, setIsTooltipVisible] = useState<boolean>(false);
-
-  const onClickImageUpload = useCallback(() => {
-    if (fileInputRef.current) fileInputRef.current.click();
-  }, []);
 
   const handleTooltip = useCallback(() => {
     setIsTooltipVisible(true);
@@ -72,23 +48,10 @@ const ModalContent = () => {
     if (isModalCommentListVisible && modalCommentImagePath.length !== 0) dispatch(modalCommentRemoveUploadedImage());
   }, [isModalCommentListVisible, modalCommentImagePath]);
 
-  const handleRemoveImage = useCallback(() => {
-    dispatch(modalCommentRemoveUploadedImage());
-  }, []);
-
   const openEditModal = useCallback(() => {
     setIsTooltipVisible(false);
     dispatch(executePostEdit());
   }, []);
-
-  const handleKeyDown = useCallback(
-    (event: React.KeyboardEvent<HTMLInputElement>) => {
-      if (event.key === 'Enter') {
-        console.log(comment);
-      }
-    },
-    [comment]
-  );
 
   return (
     <ModalContentWrapper>
@@ -146,11 +109,7 @@ const ModalContent = () => {
         </div>
       </ModalContentHeader>
 
-      {isModalCommentListVisible ? (
-        <ModalCommentList isModalCommentListVisible={isModalCommentListVisible} />
-      ) : (
-        <ModalContentText>{singlePost.content}</ModalContentText>
-      )}
+      {isModalCommentListVisible ? <ModalCommentList /> : <ModalContentText>{singlePost.content}</ModalContentText>}
 
       <ModalContentOptions $isModalCommentListVisible={isModalCommentListVisible}>
         <div>
@@ -164,48 +123,6 @@ const ModalContent = () => {
           <p>댓글 29개</p>
         </div>
       </ModalContentOptions>
-
-      <ModalCommentInputWrapper $uploading={modalCommentImagePath.length !== 0}>
-        {modalCommentImagePath.length !== 0 && (
-          <ModalCommentInputImageWrapper>
-            <ModalCommentInputImage key={modalCommentImagePath} {...slideInUploadImage}>
-              <img src={`http://localhost:3065/${modalCommentImagePath}`} alt="입력한 댓글의 첨부 이미지" />
-              <DeleteOutlined onClick={handleRemoveImage} />
-            </ModalCommentInputImage>
-          </ModalCommentInputImageWrapper>
-        )}
-
-        {isModalCommentListVisible && (
-          <ModalCommentInput $active={comment.length === 0} $uploading={modalCommentImagePath.length !== 0}>
-            <div>
-              {modalCommentUploadImageLoading ? (
-                <LoadingOutlined />
-              ) : (
-                <PaperClipOutlined onClick={onClickImageUpload} />
-              )}
-              <input
-                type="file"
-                name="image"
-                ref={fileInputRef}
-                onChange={e => onFileChange(e, modalCommentImagePath)}
-              />
-
-              <SmileOutlined />
-              <input
-                type="text"
-                placeholder="Type a Comment..."
-                value={comment}
-                onChange={onChangeComment}
-                onKeyDown={handleKeyDown}
-              />
-            </div>
-
-            <div>
-              <SendOutlined />
-            </div>
-          </ModalCommentInput>
-        )}
-      </ModalCommentInputWrapper>
     </ModalContentWrapper>
   );
 };
